add survey role authorization

+++ src/Ushahidi/Modules/V5/Http/Controllers/SurveyRoleController.php
@@ -7,6 +7,8 @@ use Illuminate\Http\Request;
 use Ushahidi\Modules\V5\Actions\Survey\Queries\FetchRolesCanCreateSurveyPostsQuery;
 use Ushahidi\Modules\V5\Actions\Survey\Commands\DeleteSurveyRolesBySurveyIDCommand;
 use Ushahidi\Modules\V5\Actions\Survey\Commands\CreateSurveyRoleCommand;
+use Ushahidi\Modules\V5\Models\Survey;
+use Ushahidi\Modules\V5\Actions\Survey\Queries\FetchSurveyByIdQuery;
 
 class SurveyRoleController extends V5Controller
 {
@@ -18,7 +20,8 @@ class SurveyRoleController extends V5Controller
      */
     public function index(int $survey_id, Request $request)
     {
-        //$this->authorizeForCurrentUser('index', SurveyRole::class);
+        // All access is based on the survey itself, not the role.
+        $this->authorize('show', new Survey());
         $resourceCollection = new SurveyRoleCollection(
             $this->queryBus->handle(
                 new FetchRolesCanCreateSurveyPostsQuery(
@@ -40,6 +43,10 @@ class SurveyRoleController extends V5Controller
      */
     public function replace(int $survey_id, Request $request)
     {
+        $survey = $this->queryBus->handle(new FetchSurveyByIdQuery($survey_id, null, null, null));
+
+        $this->authorize('update', $survey);
+
         // to do validate the role is found
         $this->commandBus->handle(new DeleteSurveyRolesBySurveyIDCommand($survey_id));
         $this->commandBus->handle(new CreateSurveyRoleCommand($survey_id, $request->input('roles')));
